wip: fetch atored tokens in airdrop setup
