+++ src/store/actions/airdropStoreActions.ts
@@ -12,8 +12,10 @@ import {
     useAirdropStore,
     useAppConfigStore,
     UserDetails,
+    UserEntryPoints,
     UserPoints,
 } from '@app/store';
+import { handleAirdropRequest } from '@app/hooks/airdrop/utils/useHandleRequest.ts';
 
 interface TokenResponse {
     exp: number;
@@ -115,6 +117,7 @@ export const airdropSetup = async () => {
         if (beConfig?.airdropUrl) {
             console.info('Refreshing airdrop tokens');
             await handleRefreshAirdropTokens(beConfig.airdropUrl);
+            await fetchAllUserData();
         }
     } catch (error) {
         console.error('Error in airdropSetup: ', error);
@@ -175,3 +178,77 @@ export const setUserGems = (userGems: number) =>
     });
 export const setUserPoints = (userPoints: UserPoints) =>
     useAirdropStore.setState({ userPoints, referralCount: userPoints?.referralCount });
+
+export const fetchAllUserData = async () => {
+    const fetchUserDetails = async () => {
+        return await handleAirdropRequest<UserDetails>({
+            path: '/user/details',
+            method: 'GET',
+            onError: handleAirdropLogout,
+        })
+            .then((data) => {
+                if (data?.user?.id) {
+                    setUserDetails(data);
+                    return data.user;
+                } else {
+                    handleAirdropLogout();
+                }
+            })
+            .catch(() => {
+                handleAirdropLogout();
+            });
+    };
+    // GET USER POINTS
+    const fetchUserPoints = async () => {
+        const data = await handleAirdropRequest<UserEntryPoints>({
+            path: '/user/score',
+            method: 'GET',
+        });
+        if (!data?.entry || !data?.entry?.gems) return;
+        setUserPoints({
+            base: {
+                gems: data.entry.gems,
+                shells: data.entry.shells,
+                hammers: data.entry.hammers,
+            },
+        });
+    };
+    // GET USER REFERRAL POINTS
+    const fetchUserReferralPoints = async () => {
+        const data = await handleAirdropRequest<{ count: ReferralCount }>({
+            path: '/miner/download/referral-count',
+            method: 'GET',
+        });
+        if (!data?.count) return;
+        setReferralCount({
+            gems: data.count.gems,
+            count: data.count.count,
+        });
+    };
+    // FETCH BONUS TIERS
+    const fetchBonusTiers = async () => {
+        const data = await handleAirdropRequest<{ tiers: BonusTier[] }>({
+            path: '/miner/download/bonus-tiers',
+            method: 'GET',
+        });
+        if (!data?.tiers) return;
+        setBonusTiers(data?.tiers);
+    };
+    const fetchData = async () => {
+        const details = await fetchUserDetails();
+
+        if (!details) return;
+        const requests: Promise<void>[] = [];
+        if (!details?.rank?.gems) {
+            requests.push(fetchUserPoints());
+        }
+        requests.push(fetchUserReferralPoints());
+        requests.push(fetchBonusTiers());
+
+        await Promise.all(requests);
+    };
+
+    if (useAirdropStore.getState().airdropTokens?.token) {
+        await fetchData();
+    }
+};
